requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
boto3>=1.39,<2.0
pillow>=11.0,<12.0
opencv-python>=4.10,<5.0
numpy>=2.0,<3.0
pydantic>=2.9,<3.0
python-multipart>=0.0.20,<1.0
httpx>=0.28,<1.0
python-dotenv>=1.0,<2.0
